Enterprise Composite Network Model
The model provides a framework for the recommended design and implementation of an enterprise campus network. The enterprise campus is comprised of the following modules:
⦁	The Building or Switch Module
⦁	The Core Module
⦁	The Management Module
⦁	The Server Module

The building or switch module is defined as the portion of the network that contains end-user workstations, phones, and their associated Layer 2 access points. Its primary goal is to provide services to end users. This module is comprised of access layer switches as well as their related distribution layer switches. This is simply the access layer in the hierarchical network model.